order detail

+++ src/main/java/com/example/shopapp/dto/OrderDetailDTO.java
@@ -21,7 +21,7 @@ public class OrderDetailDTO {
     private Long productId;
 
     @Min(value = 0, message = "Price must be greater than 0")
-    private Long price;
+    private Float price;
 
     @Min(value = 1, message = "Quantity must be greater than 0")
     @JsonProperty("number_of_products")
@@ -29,7 +29,7 @@ public class OrderDetailDTO {
 
     @Min(value = 0, message = "Total money must be greater than 0")
     @JsonProperty("total_money")
-    private int totalMoney;
+    private Float totalMoney;
 
     private String color;
 }
